Create README.md
The web-based Meeting Room Booking System is a user-friendly application which allows users to book conference rooms, manage their schedules and check availabilities. The employee and admin are the potential shareholders of this system. The user can view images of the room and detailed information about 
the venue in the user module and they can also add or update records, as well as delete their personal information from the system. By selecting user registration, the administrator can manage it for both the user and the admin. Additionally, admin can review or modify the room's information and they have access to the
booked spaces as well. To implement the system, mainly used JavaScript scripting language and Node js with Express framework used for the server-side development. HTML, CSS, JavaScript and EJS templating language used for creating different user interfaces,